requirements.txt
torch~=2.0.1+cu118
numpy~=1.25.2
scikit-learn~=1.3.0
matplotlib~=3.7.2
einops~=0.6.1
pyyaml~=6.0.1
optuna~=3.3.0
tqdm~=4.66.1
pandas~=2.0.3
pathos~=0.3.1
PyWavelets~=1.5.0
tensorboard~=2.14.0